requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pdoc==15.0.1
pluggy==1.5.0
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
pytest==8.3.5
python-gnupg==0.5.4
setuptools==78.1.0
